+++ awb_generic_sms/helpers/gateway_smart.py
@@ -1,6 +1,9 @@
 import json
 import requests
 from odoo import exceptions
+import logging
+
+_logger = logging.getLogger(__name__)
 
 
 class SmartAPIGateway(object):
@@ -38,6 +41,10 @@ class SmartAPIGateway(object):
         destination=None,
         text=None,
     ):
+        _logger.info('function: send_sms_now')
+        _logger.info(f'Mobile Number: {destination}')
+        _logger.info(f'Template Body: {text}')
+
         data = {
             'messageType': message_type,
             'destination': destination,
@@ -55,6 +62,7 @@ class SmartAPIGateway(object):
         return res
 
     def send(self):
+        _logger.info('function:send')
         sms_data = []
         for record in self.recordset:
             status_code = None
@@ -68,6 +76,8 @@ class SmartAPIGateway(object):
                 )
                 status_code = res.status_code
                 state = "sent" if status_code == 201 else "failed"
+
+                _logger.info(f'state (sent or failed): {state}')
 
             state_label = dict(self.sms_history._fields['state'].selection)[state]