+++ awb_generic_sms/docs/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [13.0.1.3.4] - 2021-09-14
+   - Welcome Notification enhancements
+   - Activation Notification enhancements
+   - Expiry Notification
+
+## [13.0.1.3.3] - 2021-09-14
+   - TMS Notification Call - to backend
+
 ## [13.0.1.3.2] - 2021-09-08
    - Aradial Welcome Notification
    - Aradial Activation Notification